+++ spinta/datasets/backends/sql/components.py
@@ -1,13 +1,13 @@
 import contextlib
 
 import sqlalchemy as sa
-
 from sqlalchemy.engine.base import Engine
 
+from spinta.components import Model
+from spinta.components import Property
+from spinta.datasets.components import ExternalBackend
 from spinta.exceptions import NoExternalName
 from spinta.exceptions import PropertyNotFound
-from spinta.components import Model, Property
-from spinta.datasets.components import ExternalBackend
 
 
 class Sql(ExternalBackend):
@@ -37,9 +37,19 @@ class Sql(ExternalBackend):
 
         return self.schema.tables[key]
 
-    def get_column(self, table: sa.Table, prop: Property, *, select=False) -> sa.Column:
+    def get_column(
+        self,
+        table: sa.Table,
+        prop: Property,
+        *,
+        select=False,
+    ) -> sa.Column:
         if prop.external is None:
             raise NoExternalName(prop)
         if prop.external.name not in table.c:
-            raise PropertyNotFound(prop.model, property=prop.name, external=prop.external.name)
+            raise PropertyNotFound(
+                prop.model,
+                property=prop.name,
+                external=prop.external.name,
+            )
         return table.c[prop.external.name]