+++ tests/datasets/test_sql.py
@@ -1,6 +1,9 @@
+import contextlib
 import pathlib
 import os
 import tempfile
+from typing import Dict
+from typing import List
 
 import pytest
 
@@ -18,27 +21,23 @@ from spinta.testing.utils import error
 from spinta.testing.client import create_remote_server
 
 
+Schema = Dict[str, List[sa.Column]]
+
+
 class Sqlite:
 
     def __init__(self, dsn: str):
         self.dsn = dsn
         self.engine = sa.create_engine(dsn)
         self.schema = sa.MetaData(self.engine)
+        self.tables = {}
 
+    def init(self, tables: Schema):
         self.tables = {
-            'country': sa.Table(
-                'salis', self.schema,
-                sa.Column('id', sa.Integer, primary_key=True),
-                sa.Column('kodas', sa.Text),
-                sa.Column('pavadinimas', sa.Text),
-            ),
-            'city': sa.Table(
-                'miestas', self.schema,
-                sa.Column('id', sa.Integer, primary_key=True),
-                sa.Column('pavadinimas', sa.Text),
-                sa.Column('salis', sa.Text),
-            )
+            k: sa.Table(k, self.schema, *v)
+            for k, v in tables.items()
         }
+        self.schema.create_all()
 
     def write(self, table, data):
         table = self.tables[table]
@@ -46,17 +45,40 @@ class Sqlite:
             conn.execute(table.insert(), data)
 
 
-@pytest.fixture(scope='module')
-def sqlite():
+@contextlib.contextmanager
+def create_sqlite_db(tables: Schema):
     with tempfile.TemporaryDirectory() as tmpdir:
         db = Sqlite('sqlite:///' + os.path.join(tmpdir, 'db.sqlite'))
-        db.schema.create_all()
-        db.write('country', [
+        db.init(tables)
+        yield db
+
+
+@pytest.fixture()
+def sqlite():
+    with tempfile.TemporaryDirectory() as tmpdir:
+        yield Sqlite('sqlite:///' + os.path.join(tmpdir, 'db.sqlite'))
+
+
+@pytest.fixture(scope='module')
+def geodb():
+    with create_sqlite_db({
+        'salis': [
+            sa.Column('id', sa.Integer, primary_key=True),
+            sa.Column('kodas', sa.Text),
+            sa.Column('pavadinimas', sa.Text),
+        ],
+        'miestas': [
+            sa.Column('id', sa.Integer, primary_key=True),
+            sa.Column('pavadinimas', sa.Text),
+            sa.Column('salis', sa.Text),
+        ],
+    }) as db:
+        db.write('salis', [
             {'kodas': 'lt', 'pavadinimas': 'Lietuva'},
             {'kodas': 'lv', 'pavadinimas': 'Latvija'},
             {'kodas': 'ee', 'pavadinimas': 'Estija'},
         ])
-        db.write('city', [
+        db.write('miestas', [
             {'salis': 'lt', 'pavadinimas': 'Vilnius'},
             {'salis': 'lv', 'pavadinimas': 'Ryga'},
             {'salis': 'ee', 'pavadinimas': 'Talinas'},
@@ -64,7 +86,7 @@ def sqlite():
         yield db
 
 
-def create_rc(rc: RawConfig, tmpdir: pathlib.Path, sqlite: Sqlite):
+def create_rc(rc: RawConfig, tmpdir: pathlib.Path, geodb: Sqlite):
     return rc.fork({
         'manifests': {
             'default': {
@@ -77,7 +99,7 @@ def create_rc(rc: RawConfig, tmpdir: pathlib.Path, sqlite: Sqlite):
         'backends': {
             'sql': {
                 'type': 'sql',
-                'dsn': sqlite.dsn,
+                'dsn': geodb.dsn,
             },
         },
         # tests/config/clients/3388ea36-4a4f-4821-900a-b574c8829d52.yml
@@ -113,13 +135,13 @@ def configure_remote_server(cli, rc: RawConfig, tmpdir: pathlib.Path, responses)
     )
 
 
-def create_client(rc: RawConfig, tmpdir: pathlib.Path, sqlite: Sqlite):
-    rc = create_rc(rc, tmpdir, sqlite)
+def create_client(rc: RawConfig, tmpdir: pathlib.Path, geodb: Sqlite):
+    rc = create_rc(rc, tmpdir, geodb)
     context = create_test_context(rc)
     return create_test_client(context)
 
 
-def test_filter(rc, tmpdir, sqlite):
+def test_filter(rc, tmpdir, geodb):
     create_tabular_manifest(tmpdir / 'manifest.csv', striptable('''
     id | d | r | b | m | property | source      | prepare   | type   | ref     | level | access | uri | title   | description
        | datasets/gov/example     |             |           |        |         |       |        |     | Example |
@@ -130,7 +152,7 @@ def test_filter(rc, tmpdir, sqlite):
        |   |   |   |   | name     | pavadinimas |           | string |         | 3     | open   |     | Name    |
     '''))
 
-    app = create_client(rc, tmpdir, sqlite)
+    app = create_client(rc, tmpdir, geodb)
 
     resp = app.get('/datasets/gov/example/country')
     assert listdata(resp, 'code', 'name') == [
@@ -138,7 +160,7 @@ def test_filter(rc, tmpdir, sqlite):
     ]
 
 
-def test_filter_join(rc, tmpdir, sqlite):
+def test_filter_join(rc, tmpdir, geodb):
     create_tabular_manifest(tmpdir / 'manifest.csv', striptable('''
     id | d | r | b | m | property | source      | prepare           | type   | ref     | level | access | uri | title   | description
        | datasets/gov/example     |             |                   |        |         |       |        |     | Example |
@@ -153,7 +175,7 @@ def test_filter_join(rc, tmpdir, sqlite):
        |   |   |   |   | country  | salis       |                   | ref    | country | 4     | open   |     | Country |
     '''))
 
-    app = create_client(rc, tmpdir, sqlite)
+    app = create_client(rc, tmpdir, geodb)
 
     resp = app.get('/datasets/gov/example/country')
     codes = dict(listdata(resp, '_id', 'code'))
@@ -166,7 +188,7 @@ def test_filter_join(rc, tmpdir, sqlite):
     ]
 
 
-def test_filter_join_array_value(rc, tmpdir, sqlite):
+def test_filter_join_array_value(rc, tmpdir, geodb):
     create_tabular_manifest(tmpdir / 'manifest.csv', striptable('''
     id | d | r | b | m | property | source      | prepare                  | type   | ref     | level | access | uri | title   | description
        | datasets/gov/example     |             |                          |        |         |       |        |     | Example |
@@ -181,7 +203,7 @@ def test_filter_join_array_value(rc, tmpdir, sqlite):
        |   |   |   |   | country  | salis       |                          | ref    | country | 4     | open   |     | Country |
     '''))
 
-    app = create_client(rc, tmpdir, sqlite)
+    app = create_client(rc, tmpdir, geodb)
 
     resp = app.get('/datasets/gov/example/country')
     codes = dict(listdata(resp, '_id', 'code'))
@@ -195,7 +217,7 @@ def test_filter_join_array_value(rc, tmpdir, sqlite):
     ]
 
 
-def test_filter_join_ne_array_value(rc, tmpdir, sqlite):
+def test_filter_join_ne_array_value(rc, tmpdir, geodb):
     create_tabular_manifest(tmpdir / 'manifest.csv', striptable('''
     id | d | r | b | m | property | source      | prepare                   | type   | ref     | level | access | uri | title   | description
        | datasets/gov/example     |             |                           |        |         |       |        |     | Example |
@@ -210,7 +232,7 @@ def test_filter_join_ne_array_value(rc, tmpdir, sqlite):
        |   |   |   |   | country  | salis       |                           | ref    | country | 4     | open   |     | Country |
     '''))
 
-    app = create_client(rc, tmpdir, sqlite)
+    app = create_client(rc, tmpdir, geodb)
 
     resp = app.get('/datasets/gov/example/country')
     codes = dict(listdata(resp, '_id', 'code'))
@@ -224,7 +246,7 @@ def test_filter_join_ne_array_value(rc, tmpdir, sqlite):
 
 
 @pytest.mark.skip('todo')
-def test_filter_multi_column_pk(rc, tmpdir, sqlite):
+def test_filter_multi_column_pk(rc, tmpdir, geodb):
     create_tabular_manifest(tmpdir / 'manifest.csv', striptable('''
     id | d | r | b | m | property | source      | prepare            | type   | ref           | level | access | uri | title   | description
        | datasets/gov/example     |             |                    |        |               |       |        |     | Example |
@@ -240,7 +262,7 @@ def test_filter_multi_column_pk(rc, tmpdir, sqlite):
        |   |   |   |   | country  | salis       |                    | ref    | country[code] | 4     | open   |     | Country |
     '''))
 
-    app = create_client(rc, tmpdir, sqlite)
+    app = create_client(rc, tmpdir, geodb)
 
     resp = app.get('/datasets/gov/example/country')
     codes = dict(listdata(resp, '_id', 'code'))
@@ -254,7 +276,7 @@ def test_filter_multi_column_pk(rc, tmpdir, sqlite):
     ]
 
 
-def test_getall(rc, tmpdir, sqlite):
+def test_getall(rc, tmpdir, geodb):
     create_tabular_manifest(tmpdir / 'manifest.csv', striptable('''
     id | d | r | b | m | property | source      | prepare | type   | ref     | level | access | uri | title   | description
        | datasets/gov/example     |             |         |        |         |       |        |     | Example |
@@ -269,7 +291,7 @@ def test_getall(rc, tmpdir, sqlite):
        |   |   |   |   | country  | salis       |         | ref    | country | 4     | open   |     | Country |
     '''))
 
-    app = create_client(rc, tmpdir, sqlite)
+    app = create_client(rc, tmpdir, geodb)
 
     resp = app.get('/datasets/gov/example/country?sort(code)')
     codes = dict(listdata(resp, '_id', 'code'))
@@ -289,7 +311,7 @@ def test_getall(rc, tmpdir, sqlite):
     ]
 
 
-def test_select(rc, tmpdir, sqlite):
+def test_select(rc, tmpdir, geodb):
     create_tabular_manifest(tmpdir / 'manifest.csv', striptable('''
     id | d | r | b | m | property | source      | prepare | type   | ref     | level | access | uri | title   | description
        | datasets/gov/example     |             |         |        |         |       |        |     | Example |
@@ -300,7 +322,7 @@ def test_select(rc, tmpdir, sqlite):
        |   |   |   |   | name     | pavadinimas |         | string |         | 3     | open   |     | Name    |
     '''))
 
-    app = create_client(rc, tmpdir, sqlite)
+    app = create_client(rc, tmpdir, geodb)
 
     resp = app.get('/datasets/gov/example/country?select(code,name)')
     assert listdata(resp, 'code', 'name') == [
@@ -311,7 +333,7 @@ def test_select(rc, tmpdir, sqlite):
 
 
 @pytest.mark.skip('TODO')
-def test_select_len(rc, tmpdir, sqlite):
+def test_select_len(rc, tmpdir, geodb):
     create_tabular_manifest(tmpdir / 'manifest.csv', striptable('''
     id | d | r | b | m | property | source      | prepare | type   | ref     | level | access | uri | title   | description
        | datasets/gov/example     |             |         |        |         |       |        |     | Example |
@@ -322,7 +344,7 @@ def test_select_len(rc, tmpdir, sqlite):
        |   |   |   |   | name     | pavadinimas |         | string |         | 3     | open   |     | Name    |
     '''))
 
-    app = create_client(rc, tmpdir, sqlite)
+    app = create_client(rc, tmpdir, geodb)
 
     resp = app.get('/datasets/gov/example/country?select(code,len(name))')
     assert listdata(resp, 'code', 'len(name)') == [
@@ -332,7 +354,7 @@ def test_select_len(rc, tmpdir, sqlite):
     ]
 
 
-def test_filter_len(rc, tmpdir, sqlite):
+def test_filter_len(rc, tmpdir, geodb):
     create_tabular_manifest(tmpdir / 'manifest.csv', striptable('''
     id | d | r | b | m | property | source      | prepare | type   | ref     | level | access | uri | title   | description
        | datasets/gov/example     |             |         |        |         |       |        |     | Example |
@@ -343,7 +365,7 @@ def test_filter_len(rc, tmpdir, sqlite):
        |   |   |   |   | name     | pavadinimas |         | string |         | 3     | open   |     | Name    |
     '''))
 
-    app = create_client(rc, tmpdir, sqlite)
+    app = create_client(rc, tmpdir, geodb)
 
     resp = app.get('/datasets/gov/example/country?select(code,name)&len(name)=7&sort(code)')
     assert listdata(resp, 'code', 'name') == [
@@ -352,7 +374,7 @@ def test_filter_len(rc, tmpdir, sqlite):
     ]
 
 
-def test_private_property(rc, tmpdir, sqlite):
+def test_private_property(rc, tmpdir, geodb):
     create_tabular_manifest(tmpdir / 'manifest.csv', striptable('''
     id | d | r | b | m | property | source      | prepare    | type   | ref     | level | access  | uri | title   | description
        | datasets/gov/example     |             |            |        |         |       |         |     | Example |
@@ -363,7 +385,7 @@ def test_private_property(rc, tmpdir, sqlite):
        |   |   |   |   | name     | pavadinimas |            | string |         | 3     | open    |     | Name    |
     '''))
 
-    app = create_client(rc, tmpdir, sqlite)
+    app = create_client(rc, tmpdir, geodb)
 
     resp = app.get('/datasets/gov/example/country')
     assert listdata(resp) == [
@@ -372,7 +394,7 @@ def test_private_property(rc, tmpdir, sqlite):
     ]
 
 
-def test_all_private_properties(rc, tmpdir, sqlite):
+def test_all_private_properties(rc, tmpdir, geodb):
     create_tabular_manifest(tmpdir / 'manifest.csv', striptable('''
     id | d | r | b | m | property | source      | prepare    | type   | ref     | level | access  | uri | title   | description
        | datasets/gov/example     |             |            |        |         |       |         |     | Example |
@@ -383,13 +405,13 @@ def test_all_private_properties(rc, tmpdir, sqlite):
        |   |   |   |   | name     | pavadinimas |            | string |         | 3     | private |     | Name    |
     '''))
 
-    app = create_client(rc, tmpdir, sqlite)
+    app = create_client(rc, tmpdir, geodb)
 
     resp = app.get('/datasets/gov/example/country')
     assert error(resp, status=401) == 'AuthorizedClientsOnly'
 
 
-def test_default_access(rc, tmpdir, sqlite):
+def test_default_access(rc, tmpdir, geodb):
     create_tabular_manifest(tmpdir / 'manifest.csv', striptable('''
     id | d | r | b | m | property | source      | prepare    | type   | ref     | level | access  | uri | title   | description
        | datasets/gov/example     |             |            |        |         |       |         |     | Example |
@@ -400,13 +422,13 @@ def test_default_access(rc, tmpdir, sqlite):
        |   |   |   |   | name     | pavadinimas |            | string |         | 3     |         |     | Name    |
     '''))
 
-    app = create_client(rc, tmpdir, sqlite)
+    app = create_client(rc, tmpdir, geodb)
 
     resp = app.get('/datasets/gov/example/country')
     assert error(resp, status=401) == 'AuthorizedClientsOnly'
 
 
-def test_model_open_access(rc, tmpdir, sqlite):
+def test_model_open_access(rc, tmpdir, geodb):
     create_tabular_manifest(tmpdir / 'manifest.csv', striptable('''
     id | d | r | b | m | property | source      | prepare    | type   | ref     | level | access  | uri | title   | description
        | datasets/gov/example     |             |            |        |         |       |         |     | Example |
@@ -417,7 +439,7 @@ def test_model_open_access(rc, tmpdir, sqlite):
        |   |   |   |   | name     | pavadinimas |            | string |         | 3     |         |     | Name    |
     '''))
 
-    app = create_client(rc, tmpdir, sqlite)
+    app = create_client(rc, tmpdir, geodb)
 
     resp = app.get('/datasets/gov/example/country')
     assert listdata(resp) == [
@@ -426,7 +448,7 @@ def test_model_open_access(rc, tmpdir, sqlite):
     ]
 
 
-def test_property_public_access(rc, tmpdir, sqlite):
+def test_property_public_access(rc, tmpdir, geodb):
     create_tabular_manifest(tmpdir / 'manifest.csv', striptable('''
     id | d | r | b | m | property | source      | prepare    | type   | ref     | level | access  | uri | title   | description
        | datasets/gov/example     |             |            |        |         |       |         |     | Example |
@@ -437,7 +459,7 @@ def test_property_public_access(rc, tmpdir, sqlite):
        |   |   |   |   | name     | pavadinimas |            | string |         | 3     | open    |     | Name    |
     '''))
 
-    app = create_client(rc, tmpdir, sqlite)
+    app = create_client(rc, tmpdir, geodb)
 
     resp = app.get('/datasets/gov/example/country')
     assert listdata(resp) == [
@@ -452,7 +474,7 @@ def test_property_public_access(rc, tmpdir, sqlite):
     ]
 
 
-def test_select_protected_property(rc, tmpdir, sqlite):
+def test_select_protected_property(rc, tmpdir, geodb):
     create_tabular_manifest(tmpdir / 'manifest.csv', striptable('''
     id | d | r | b | m | property | source      | prepare    | type   | ref     | level | access  | uri | title   | description
        | datasets/gov/example     |             |            |        |         |       |         |     | Example |
@@ -463,7 +485,7 @@ def test_select_protected_property(rc, tmpdir, sqlite):
        |   |   |   |   | name     | pavadinimas |            | string |         | 3     | open    |     | Name    |
     '''))
 
-    app = create_client(rc, tmpdir, sqlite)
+    app = create_client(rc, tmpdir, geodb)
 
     resp = app.get('/datasets/gov/example/country?select(code,name)')
     assert error(resp) == 'PropertyNotFound'
@@ -472,7 +494,7 @@ def test_select_protected_property(rc, tmpdir, sqlite):
     assert error(resp) == 'PropertyNotFound'
 
 
-def test_ns_getall(rc, tmpdir, sqlite):
+def test_ns_getall(rc, tmpdir, geodb):
     create_tabular_manifest(tmpdir / 'manifest.csv', striptable('''
     id | d | r | b | m | property | source      | prepare    | type   | ref     | level | access  | uri | title   | description
        | datasets/gov/example     |             |            |        |         |       |         |     | Example |
@@ -483,7 +505,7 @@ def test_ns_getall(rc, tmpdir, sqlite):
        |   |   |   |   | name     | pavadinimas |            | string |         | 3     | open    |     | Name    |
     '''))
 
-    app = create_client(rc, tmpdir, sqlite)
+    app = create_client(rc, tmpdir, geodb)
 
     resp = app.get('/datasets/gov/example')
     assert listdata(resp, '_id', 'title') == [
@@ -496,7 +518,7 @@ def test_ns_getall(rc, tmpdir, sqlite):
     ]
 
 
-def test_push(postgresql, rc, cli, responses, tmpdir, sqlite, request):
+def test_push(postgresql, rc, cli, responses, tmpdir, geodb, request):
     create_tabular_manifest(tmpdir / 'manifest.csv', striptable('''
     d | r | b | m | property | source      | type   | ref     | access
     datasets/gov/example     |             |        |         |
@@ -516,7 +538,7 @@ def test_push(postgresql, rc, cli, responses, tmpdir, sqlite, request):
     request.addfinalizer(remote.app.context.wipe_all)
 
     # Configure local server with SQL backend
-    localrc = create_rc(rc, tmpdir, sqlite)
+    localrc = create_rc(rc, tmpdir, geodb)
 
     # Push data from local to remote.
     cli.invoke(localrc, push, [
@@ -546,7 +568,7 @@ def test_push(postgresql, rc, cli, responses, tmpdir, sqlite, request):
     ]
 
     # Add new data to local server
-    sqlite.write('city', [
+    geodb.write('miestas', [
         {'salis': 'lt', 'pavadinimas': 'Kaunas'},
     ])
 
@@ -576,7 +598,7 @@ def test_push(postgresql, rc, cli, responses, tmpdir, sqlite, request):
     ]
 
 
-def test_no_primary_key(rc, tmpdir, sqlite):
+def test_no_primary_key(rc, tmpdir, geodb):
     create_tabular_manifest(tmpdir / 'manifest.csv', striptable('''
     d | r | b | m | property | source      | type   | ref | access
     datasets/gov/example     |             |        |     |
@@ -587,7 +609,7 @@ def test_no_primary_key(rc, tmpdir, sqlite):
       |   |   |   | name     | pavadinimas | string |     | open
     '''))
 
-    app = create_client(rc, tmpdir, sqlite)
+    app = create_client(rc, tmpdir, geodb)
 
     resp = app.get('/datasets/gov/example/country')
     codes = dict(listdata(resp, '_id', 'code'))
@@ -600,7 +622,7 @@ def test_no_primary_key(rc, tmpdir, sqlite):
     ]
 
 
-def test_count(rc, tmpdir, sqlite):
+def test_count(rc, tmpdir, geodb):
     create_tabular_manifest(tmpdir / 'manifest.csv', striptable('''
     d | r | b | m | property | source      | type   | ref | access
     datasets/gov/example     |             |        |     |
@@ -611,13 +633,13 @@ def test_count(rc, tmpdir, sqlite):
       |   |   |   | name     | pavadinimas | string |     | open
     '''))
 
-    app = create_client(rc, tmpdir, sqlite)
+    app = create_client(rc, tmpdir, geodb)
 
-    resp = app.get('/datasets/gov/example/country?count()')
+    resp = app.get('/datasets/gov/example/country?select(count())')
     assert listdata(resp) == [3]
 
 
-def test_push_chunks(postgresql, rc, cli, responses, tmpdir, sqlite, request):
+def test_push_chunks(postgresql, rc, cli, responses, tmpdir, geodb, request):
     create_tabular_manifest(tmpdir / 'manifest.csv', striptable('''
     d | r | b | m | property | source      | type   | ref     | access
     datasets/gov/example     |             |        |         |
@@ -633,7 +655,7 @@ def test_push_chunks(postgresql, rc, cli, responses, tmpdir, sqlite, request):
     request.addfinalizer(remote.app.context.wipe_all)
 
     # Configure local server with SQL backend
-    localrc = create_rc(rc, tmpdir, sqlite)
+    localrc = create_rc(rc, tmpdir, geodb)
 
     # Push data from local to remote.
     cli.invoke(localrc, push, [
@@ -653,7 +675,7 @@ def test_push_chunks(postgresql, rc, cli, responses, tmpdir, sqlite, request):
     ]
 
 
-def test_push_state(postgresql, rc, cli, responses, tmpdir, sqlite, request):
+def test_push_state(postgresql, rc, cli, responses, tmpdir, geodb, request):
     create_tabular_manifest(tmpdir / 'manifest.csv', striptable('''
     d | r | b | m | property | source      | type   | ref     | access
     datasets/gov/example     |             |        |         |
@@ -669,7 +691,7 @@ def test_push_state(postgresql, rc, cli, responses, tmpdir, sqlite, request):
     request.addfinalizer(remote.app.context.wipe_all)
 
     # Configure local server with SQL backend
-    localrc = create_rc(rc, tmpdir, sqlite)
+    localrc = create_rc(rc, tmpdir, geodb)
 
     # Push one row, save state and stop.
     cli.invoke(localrc, push, [
@@ -700,7 +722,7 @@ def test_push_state(postgresql, rc, cli, responses, tmpdir, sqlite, request):
     assert len(listdata(resp)) == 2
 
 
-def test_prepared_property(rc, tmpdir, sqlite):
+def test_prepared_property(rc, tmpdir, geodb):
     create_tabular_manifest(tmpdir / 'manifest.csv', striptable('''
     d | r | b | m | property  | source      | prepare   | type   | ref  | access
     datasets/gov/example      |             |           |        |      |
@@ -711,11 +733,65 @@ def test_prepared_property(rc, tmpdir, sqlite):
       |   |   |   | continent |             | 'EU'      | string |      |
     '''))
 
-    app = create_client(rc, tmpdir, sqlite)
+    app = create_client(rc, tmpdir, geodb)
 
     resp = app.get('/datasets/gov/example/country')
     assert listdata(resp, 'continent', 'code', 'name') == [
         ('EU', 'ee', 'Estija'),
         ('EU', 'lt', 'Lietuva'),
         ('EU', 'lv', 'Latvija'),
+    ]
+
+
+def test_composite_keys(rc, tmpdir, sqlite):
+    create_tabular_manifest(tmpdir / 'manifest.csv', striptable('''
+    d | r | m | property | source | prepare | type   | ref     | access
+    datasets/ds          |        |         |        |         |
+      | rs               |        |         | sql    |         |
+      |   | t1           | t1     |         |        | a1      | open
+      |   |   | p1       | p1     |         | string |         |
+      |   |   | p2       | p2     |         | string |         |
+      |   |   | a1       |        | p1, p2  | ref    | t2      |
+      |   | t2           | t2     |         |        | p3, p4  | open
+      |   |   | p3       | p3     |         | string |         |
+      |   |   | p4       | p4     |         | string |         |
+    '''))
+
+    app = create_client(rc, tmpdir, sqlite)
+
+    sqlite.init({
+        't1': [
+            sa.Column('p1', sa.Text),
+            sa.Column('p2', sa.Text),
+        ],
+        't2': [
+            sa.Column('p3', sa.Text),
+            sa.Column('p4', sa.Text),
+        ],
+    })
+
+    sqlite.write('t1', [{'p1': 'a', 'p2': 'b'}])
+    sqlite.write('t2', [{'p3': 'a', 'p4': 'b'}])
+
+    resp = app.get('/datasets/ds/t2')
+    data = listdata(resp, '_id', 'p3', 'p4', sort='p3')
+    t2pk = {_id: (p3, p4) for _id, p3, p4 in data}
+    data = [(t2pk.get(_id), p3, p4) for _id, p3, p4 in data]
+    assert data == [(('a', 'b'), 'a', 'b')]
+
+    resp = app.get('/datasets/ds/t1')
+    data = listdata(resp, 'p1', 'p2', 'a1', sort='p1')
+    data = [
+        (
+            p1,
+            p2,
+            {
+                **a1,
+                '_id': t2pk.get(a1.get('_id')),
+            },
+        )
+        for p1, p2, a1 in data
+    ]
+    assert data == [
+        ('a', 'b', {'_id': ('a', 'b')}),
     ]