+++ tests/test_exceptions.py
@@ -203,5 +203,6 @@ def test_this_dataset_model_property(context):
         '    model: datasets/backends/postgres/dataset/report\n'
         '    entity: reports\n'
         '    property: status\n'
+        '    attribute: None\n'
         '    resource.backend: default\n'
     )